requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0
pydantic==2.5.3
python-dotenv==1.0.1
httpx==0.26.0
pytest==8.0.0
email-validator==2.1.0.post1